chore: wip refactoring tests
