+++ tests/unit/test_public_image_checks.py
@@ -1,10 +1,4 @@
-import pytest
-
-from renku_notebooks.util.check_image import (
-    parse_image_name,
-    image_exists,
-    get_docker_token,
-)
+from renku_notebooks.util.check_image import parse_image_name
 
 
 def test_public_image_name_parsing():
@@ -123,21 +117,3 @@ def test_public_image_name_parsing():
         "image": "proj/image/subimage",
         "tag": "latest",
     }
-
-
-@pytest.mark.integration
-def test_public_image_check():
-    parsed_image = parse_image_name("nginx:1.19.3")
-    token, _ = get_docker_token(**parsed_image, user={})
-    assert image_exists(**parse_image_name("nginx:1.19.3"), token=token)
-    parsed_image = parse_image_name("nginx")
-    token, _ = get_docker_token(**parsed_image, user={})
-    assert image_exists(**parse_image_name("nginx"), token=token)
-    parsed_image = parse_image_name("renku/singleuser:cb70d7e")
-    token, _ = get_docker_token(**parsed_image, user={})
-    assert image_exists(**parsed_image, token=token)
-    parsed_image = parse_image_name("renku/singleuser")
-    token, _ = get_docker_token(**parsed_image, user={})
-    assert image_exists(**parsed_image, token=token)
-    parsed_image = parse_image_name("madeuprepo/madeupproject:tag")
-    assert not image_exists(**parsed_image, token="madeuptoken")
